refactor: removes duplicate code
