+++ services/hash.py
